requirements.txt
python-telegram-bot==21.1.1
google-api-python-client==2.149.0
google-auth==2.36.0
google-auth-oauthlib==1.2.0
flask==3.0.2
apscheduler==3.10.4
